Wifi: The connection problem
I have added a wifi connecting function prototype. However, a problem I have is with the wifi passwords, I have no way to  detect if the connection was made reliably and the connection just never works in the first place. I need a better way to connect to the wifi or fix my broken code.