Kumina: I have started work on a kumina scripting service
I have created and edited the kumina program as a whole to be able to support scripting.  The programming language will have to be a scripting language because of the limitations. It will only be for automation. Also user must have a udr activation. Script will be compiled for security issues. Find a way to store a long string within the base 85